add formatted printing of QObjects
This patch adds a function to format a QDict according to a string
that is printf like but specifies how to navigate the QDict.  All
parts are printed as JSON, but this is fine since this will auto-escape
strings and integers print naturally.

More sophisticated flags can be added later, e.g. %#{foo} would
print a string without escaping.

Cc: Luiz Capitulino <lcapitulino@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>