malloc: add statistics infrastructure
The infrastructure tracks the following items for each allocation locus:
number of currently allocated blocks, total number of allocated blocks,
number of currently allocated bytes, maximum number of allocated bytes,
total number of allocated bytes.  Accounting is thread safe, which is
why I didn't use qemu-queue functions.

Tracing is done by sticking an extra header in front of the allocated
block.  The header is needed so that free can account bytes correctly
and to the correct allocation locus.  The header is variable-sized
because of memalign, with the information stored at the tail so it is
easily accessible from free's parameter.  Hence it also has to include
a pointer to the actual base of the memory block.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>