omap_clk: convert to QOM
Make the OMAP clocks QOM objects, so that they we can put them in a branch
of the object graph and turn the pointer properties into links.

As in the current code, the clocks are copied into a freshly allocated
memory block when they are initialized.  However, the "templates" now
have a separate struct definition because the header must be created
anyway with object_initialize.  The linking of children to parents
is moved to a separate loop for simplicity.

Cc: Peter Maydell <peter.maydell@linaro.org>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>