qom: revamp interfaces
This commit is composed of several changes:

1) it provides a complete set of macros that serve as building blocks
for the definition of Interface types.

2) make Interface public.  This is because this type serves as the
base class for the actual implementation objects that are returned
by dynamic_cast.  While the type itself need not be subclassed by
clients and in principle could remain opaque, it is useful in helper
macros.

3) Change the definition of Interface, prefixing field names with
"iface_" so that the macros can perform a limited a amount of static
type checking.

4) Add documentation for interfaces.

It doesn't change the interface SList to a type-safe qemu-queue.h
list.  This can be left for later.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>