raw-posix: keep complete control of door locking if possible
Try to open the disk O_EXCL so that udev will not receive eject request
and media change events.  If we cannot do that, and the guest is using
scsi-block, the host and guest would race on receiving media change
notifications from the disc.  So, without O_EXCL the only sensible
choice is to ignore completely the guest's door management.

With IDE passthrough, we do not need O_EXCL in principle.  However,
using the emulated passthrough CD-ROM gets confusing without O_EXCL
(to the user and to the guest, always; and sometimes, to the host udev
as well).  The guest tries to lock the disk, but the host udev runs as
root and will unlock it as soon as it receives an eject request.

This patch does the bulk of the work.  For SCSI to work we need to
try harder when open(O_EXCL) fails, which we do in the next patches by
talking to udisks.  Fixing IDE also requires tracking the host tray state
and converting it to media change events and sense codes in the guest.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>