ide: derive ascq from tray_open
The sense codes that QEMU is providing are not too helpful.  We can
provide information on the tray status in the additional sense code
qualified.  Luckily, that's all that we need to do with the ascq,
so we can build it on the fly when processing REQUEST SENSE and not
care about migration.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>