ide: track host tray state
If we find that the host's tray changed state, report this to the guest.
The change media callback is perfect for this!

With this change, the tray state will always be "closed" when the
media state is MEDIUM_OK, so the conditionals for the cdrom_changed
state machine can be simplified further.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>