ide: report media readiness
If the host says that the medium is not ready (but "is in the process
of becoming ready", as SPC puts it) the NOT READY sense key is fine,
but we should report a different asc/ascq pair.

Also, report the "medium not present" unit attention condition as soon
as the new medium has been detected, but delay "medium may have
changed" until after the drive leaves not ready state.  This has
no effect on operation with ISO images, because they do not report
MEDIUM_NOT_READY, but matches this sequence of sense codes from a
real drive:

1) Sense key: Not Ready, Additional sense: Medium not present - tray open
2) Sense key: Not Ready, Additional sense: Medium not present - tray closed
3) Sense key: Not Ready, Additional sense: Logical unit is in process
   of becoming ready
4) Sense key: Unit Attention, Additional sense: Not ready to ready change,
   medium may have changed

(The sense codes were obtained with a continuous stream of TEST UNIT READY
commands, using sg_turs while the CD is opened with O_EXCL):

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>