raw-posix: do not report tray movement errors
Emulation of the CD-ROM tray has to be cooperative; you can never
be sure that nobody else opens /dev/cdrom after you, and at that
point the kernel will prevent you from ejecting the disk.  So we
cannot be sure that the ioctls will succeed, and printing errors
with strerror is not going to produce helpful messages in general;
better shut up and swallow.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>