block: poll for host eject requests
We would like the host eject button to work even when the host CD drive
is being passed down as an emulated IDE drive.  To do this, whenever the
guest sends GET EVENT STATUS NOTIFICATION we look for eject request events
and pass them down.

This is a major usability improvement, so we do it for emulated SCSI disks
as well (even though for SCSI it is simpler and better to use scsi-block
instead).

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>