qemu-timer: make thread-safe
This is pretty simple; it uses a mutex in the QEMUClock.  There are
two interesting points:

1) rearming the timer must be done under the iothread mutex.
We use a bottom half for this;

2) to avoid excessive locking/unlocking, the deadline is stored in
the QEMUClock before unlocking the mutex and triggering the bottom
half.  The bottom half also ensures that qemu_next_alarm_deadline can
proceed with an obsolete deadline, because in that case the iothread's
select() will wake up immediately.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>