add generic thread-pool facility
Add a generic thread-pool.  The code is roughly based on posix-aio-compat.c,
with some changes.  The main is to use QemuSemaphore instead of QemuCond;
another is to separate the state of the thread from the return code of the
worker function.  The return code is totally opaque for the thread pool.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>