add dependent wait
Allow a thread to wait until another work item has completed.
This cannot be done simply with a condition variable because we
do not want to allocate synchronization primitives per work item
(only per-thread).

So we use a single condvar to detect threads moving out of
THREAD_QUEUED.  Once they do, they will have a thread assigned
to them and we can wait on that thread's own condvar.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>