use normal thread pool for raw
This patch moves the I/O logic from posix-aio-compat.c to raw-posix.c.
Instead of calling into posix-aio-compat.c with the asynchronous methods,
it simply provides synchronous I/O methods that are executed in the
thread pool.

Asynchronous I/O methods are still provided, giving a fast path to call
into linux-aio.c.  If they returns NULL the thread pool will be used.
This means that linux-aio will not be used neither for copy-on-read,
nor for formats such as qcow2.  This is okay, since in both cases
the (raw) I/O was being done synchronously.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>