block: do not stream sectors that are unallocated in the base
Full streaming of a block device will always populate all clusters of the
destination, with logical zero clusters in the case of qed but with full
clusters in the case of qcow2.  However, this is not necessary; even
with qed it is bad because it will perform useless reads of the source,
which could be on a slow network medium).  A cluster that is unallocated
should never be read by the guests, except for silly things such as
"dd if=/dev/vda of=/dev/null", so it is not really necessary to stream
it.  If the guest does those silly things we will have to read it only
once from the slow source, since the behavior does not change for
guest-initiated copy-on-read.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>