blkdebug: store list of active rules
This prepares for the next patch, where some active rules may actually
not trigger depending on input to readv/writev.  Store the active rules
in a SIMPLEQ (so that it can be emptied easily with QSIMPLEQ_INIT), and
fetch the errno/once/immediately arguments from there.

There is a small behavioral change.  If two rules exist for the same
state and event, and one of them is disabled with once=true, the enabled
one will be enforced.  With the previous code, the last rule would always
win and (if disabled) it would disable all previous rules.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>