block: return computed BDRV_O_CACHE_WB from bdrv_get_flags
I include in this submission an eighth patch, that modifies Markus's
recently posted bdrv_get_flags cache function to include the computed
value of BDRV_O_CACHE_WB.  I'm not sure if this is a good idea or not,
it just happened to me while reviewing Markus's patches.  On one hand
it makes sense because this is the only similarly mutable flag.  On the
other, bdrv_enable_write_cache() already provides the same info; also,
flags such as BDRV_O_SNAPSHOT and BDRV_O_NO_BACKING are anyway removed
by bdrv_get_flags, so the function should not be used to reopen the
file---at least not as it is now.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>