scsi: report TAS=0, do not report TASK ABORTED
Another mode value that differs under emulation is TAS, which is 1 if
commands aborted from another I_T nexus are completed with a
TASK_ABORTED status.  We usually do not care, because we implement
a single I_T nexus.

However, TASK ABORTED status might happen if we do SCSI passthrough, and
we want to change it to a cancellation so that we "enforce" TAS=0.

This unfortunately makes for an ugly API name.  scsi_req_abort should
really be called scsi_req_cancel (since it cancels I/O); scsi_req_cancel
should really be called scsi_req_abort (since it aborts a command).

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>