exec: small adjustments for TCG separation
Add stubs for cpu_restore_state and tlb_fill, which should respectively
have no effect and never be used outside TCG mode.

Add assertions that TCG is enabled around code that calls to TCG from
exec.c, so that the compiler can remove the entire functions as dead
(the proper way to do this should be to move them to cpu-exec.c
instead).

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>