qemu-thread: introduce qemu_mutex_lock_recursive
This function introduces a function to lock mutexes recursively.
Recursive mutexes are not portable when used in combination with condition
variables, so this introduces a different function for recursive locking,
to be used like this

    int recursive = qemu_mutex_lock_recursive(&mutex);
    ...
    if (!recursive) {
        qemu_mutex_unlock(&mutex);
    }

Usually the recursion will not be needed and it will be able to use
condition variables just fine.  When recursion is needed, the above
idiom will make it clear.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>