qdev: switch reset to post-order
Post-order is the only sensible direction for the reset signals.  For example,
suppose pre-order is used and the parent has some data structures that cache
children state (for example a list of active requests).  When the reset
method is invoked on the parent, these caches could be in any state.

If post-order is used, on the other hand, these will be in a known state
when the reset method is invoked on the parent.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>