qdev: do not reset a device until the parent has been initialized
When a device creates a bus and more devices as part of its init callback, the
child device could be reset while the parent is still only partly initialized.
In this case, the right thing to do is to delay resetting the child.  Do not
do it at all in qdev_init, instead use qdev_reset_all to reset already-created
devices when the state goes from CREATED to INITIALIZED.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>