kvm: reset state from the CPU's reset method
Now that we have a CPU object with a reset method, it is better to
keep the KVM reset close to the CPU reset.  Using qemu_register_reset
as we do now keeps them far apart.

With this patch, ARM and PPC no longer call the kvm_arch_ function, so
it can get removed there.  i386 and s390x call it from X86CPU/S390CPU
reset handler, and the function gets an X86CPU/S390CPU.

Note that s390-specific functions are called kvm_s390_*, while
x86-specific functions are called kvm_arch_*.  Changing that is the
topic of a separate patch.

Reviewed-by: Gleb Natapov <gnatapov@redhat.com>
Reviewed-by: Andreas Färber <afaerber@suse.de>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>