exec: make iotlb RCU-friendly
After the previous patch, TLBs will synchronize on every change to
the memory mapping.  This patch augments that with synchronization
of the MemoryRegionSections cached in the iotlb array.

Every change to RCU-protected data can now force an exit of the
current translation block, and the new data structure is updated
in the CPUArchState, so that each CPU_INTERRUPT_TLBFLUSH interrupt
constitutes the boundary of an RCU read-side critical section.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>