qemu-char: do not use error_is_set
error_is_set must only be used if, for some reason, you want to elide a whole
function upon receiving a non-NULL error.  This is never the case in qemu-char,
so we must use a local Error variable and propagate it to errp.

Also works towards fixing some resource leaks reported by Coverity (where a
file descriptor is opened, but then leaked if *errp happened to be non-NULL:
this would be a case where we should have used error_is_set *in the callee*,
while this patch removes uses in the caller).

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>