vga: sort-of implement word and double-word access modes
For some weird reason, Jazz Jackrabbit uses odd/even mode with 256-color
graphics, with byte mode instead of word mode as is the case in text
modes.

Implementing the modes properly is a mess in QEMU, because it would
change the layout of VRAM and break migration.  As an approximation,
shift right when the CPU accesses memory instead of shifting left
when the CRT controller reads it.  A hack is needed in order to
write font data properly (see comment in the code), but it works
well enough for the game.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>