vga: use latches in odd/even mode too
For some weird reason, Jazz Jackrabbit uses odd/even mode with 256-color
graphics.  Accesses still use all planes (reads go to the latches and
the game uses read mode 1 so that the CPU always gets 0xFF; writes use
the plane mask register because the game sets bit 2 of the sequencer's
memory mode register).

This patch thus uses the code for latched memory accesses in odd/even
mode as well.  The only difference is how the plane is computed in
read mode 0, and how the planes are masked if the aforementioned bit
2 is reset.

It is almost enough to fix the game.  You also need to honor byte/word
mode selection, which is done in the next patch.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>