pci-assign: munmap the BARs in instance_finalize
Because the BARs are not MMIO memory regions, they will be accessed outside
the big QEMU lock by default.  The memory core will take care of keeping
the device alive until the end of the RCU grace period, because the
FlatViews take a reference the device when they do memory_region_ref
on the BARs themselves.  However, the unrealize callback is too early:
the munmap must be delayed until the last reference to the device goes
away.

To do this, split free_assigned_device in two.  The removal step is still
called by assigned_exitfn.  The reclamation step is unmap_assigned_device
and is moved to the instance_finalize callback.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>