vfio: munmap the BARs in instance_finalize
Because the BARs are not MMIO memory regions, they will be accessed outside
the big QEMU lock by default.  The memory core will take care of keeping
the device alive until the end of the RCU grace period, because the
FlatViews take a reference the device when they do memory_region_ref
on the BARs themselves.  However, the unrealize callback is too early:
the munmap must be delayed until the last reference to the device goes
away.

To do this, split vfio_unmap_bars in two.  The removal step, now called
vfio_unregister_bars, remains in vfio_exitfn.  The reclamation step
is vfio_unmap_bars and is moved to the instance_finalize callback.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>