exec: use memory_region_is_logging to optimize dirty tracking
memory_region_is_logging now returns the exact set of bitmaps that
have to be tracked.  Use it instead of the in_migration variable.

In the next patch, we will also use it to set only DIRTY_MEMORY_VGA
or DIRTY_MEMORY_MIGRATION if necessary.  As a result, fewer expensive
atomic operations will be required.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>