exec: only check relevant bitmaps for cleanliness
Most of the time, not all bitmaps have to be marked as dirty;
do not do anything if the interesting ones are already dirty.
Previously, any clean bitmap would have cause all the bitmaps to be
marked dirty.

In fact, unless running TCG most of the time bitmap operations
are not done at all because memory_region_is_logging returns
zero.  In this case, skip cpu_physical_memory_range_includes_clean
as well.

With this patch, cpu_physical_memory_set_dirty_range is called
unconditionally, so there need not be anymore a separate call to
xen_modified_memory.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>