vga: disable chain4_alias if KVM supports SMRAM
KVM is okay with SMRAM overlapping an MMIO area underneath it, but it does
not want to have two overlapping RAM slots for SMRAM and video RAM.
Unfortunately, the chain4_alias optimization results in the latter
situation.  Disable it if KVM supports system management mode.

Note that the chain4_alias optimization is misguided, because it assumes
that chain4 data is at VRAM address 0,1,2,3,4,5,6,7...16383.  This is
incorrect, because chain4 data is at VRAM address 0,1,2,3, 16,17,18,19,
..., 65523.  But we cannot fix the VRAM format without breaking migration,
so keep the optimization and just disable it if it gets in the way.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>