memory: add kvm_mem_flags to MemoryRegion
This patch is ugly; it adds a KVM-specific flag to memory regions that
is used to distinguish SMRAM regions from others.  This of course is a
layering violation, but I have no other good ideas about how to avoid it.

If you let KVM use address_space_memory as it did until now, and add
separate calls to KVM_SET_USER_MEMORY_REGION for SMRAM regions, it's
much harder to cover the case when a RAM or ROM BAR is placed to cover
parts of SMRAM.

Hence, this series uses instead one address space only, including
all of system memory, PCI BARs and SMRAM.  This way, hiding the BARs
is done automatically by the memory core.  This, however, has an
important issue which requires this kind of layering violation.

The issue is that when you close SMRAM there is no change in the memory
map as far as the memory API is concerned.  The only change is in the
flags we pass to KVM, but these are not visible to the memory API.  So
KVM needs to convince the memory API to see a region_del/region_add
sequence from the listener; it needs to convince flatrange_equal to
return false for the "open SMRAM" and "closed SMRAM" cases.  This is
what the kvm_mem_flags do.

Besides the layering violation, there is also the question of how to
handle flags from subregions.  The current code just ORs the parent
region's flags into all subregions.

This being ugly, and hopefully temporary until someone shows me the
light, I haven't yet added nice accessors for the field.  The final
patch accesses the field blindly from target-i386/kvm.c.

One alternative would be to point two different RAM regions to the same
guest memory, using memory_region_init_ram_ptr.  That works, but it's
even more gross.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>