aio: introduce aio_poll_internal
Move the implemntation of aio_poll to aio_poll_internal, and introduce
a wrapper for public use.  For now it just asserts that aio_poll is
being used correctly---either from the thread that manages the context,
or with the QEMU global mutex held.

The next patch, however, will completely separate the two cases.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>