aaync: avoid blocking poll
Non-blocking waits can be a little more efficient.  Use two tricks in
order to make them more frequent.

First, if ctx->notified is true we know we have a ready file
descriptor.  In this case, there is no need to do a blocking wait
and thus the ctx->notify_me field can always stay false.  This
saves on event_notifier_set calls, and applies to both POSIX and
Win32.

Second, we want to delay aio_notify_accept as much as possible, possibly
forever if the iothread is consuming all of its CPU share.  Therefore,
if more than one file descriptor is readable, we do not call
aio_notify_accept.  This does not reset ctx->notified is false; the
next iteration will also be non-blocking and will also economize on
event_notifier_set calls.  In the end, if ctx->notified was false the
number of readable file descriptors will fall to zero; if it was true
eventually the context's EventNotifier will be the only readable
file descriptor and aio_notify_accept will be called.

This only applies to POSIX, because Windows uses a different API that
only returns one event at a time.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>