virtio: read/write the VirtQueueElement a field at a time
This is how it should have done from the beginning, but unfortunately
it wasn't noticed in commit 869a5c6 ("Stop VM on error in virtio-blk",
2009-01-22) and later propagated from virtio-blk to virtio-serial and
virtio-scsi.

This breaks migration compatibility *in both directions* for 32-bit
builds, and backwards for 64-bit builds.

For virtio-blk and virtio-scsi it's only in the rare case where migration
is done while the VM is in I/O error state, and for virtio-serial in
the (relatively more common) case where the port was throttled on the
source machine.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>