cpus-common: remove redundant call to exclusive_idle()
No need to call exclusive_idle() from cpu_exec_end since it is done
immediately afterwards in cpu_exec_start.  Any exclusive section could
run as soon as cpu_exec_end because cpu->running is false and the mutex
is not taken, so this does not add any protection either.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>