block/snapshot: do not take AioContext lock
Snapshots are only created/destroyed/loaded under the BQL, while no
other I/O is happening.  Snapshot information could be accessed while
other I/O is happening, but also under the BQL so they cannot be
modified concurrently.  The AioContext lock is overkill.  If needed,
in the future the create/destroy/goto callbacks could be changed
to run in a coroutine, so the driver can take care of mutual exclusion
as usual, and the BQL could be split to a separate lock covering all
snapshot operations.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>