dirty-bitmap: ensure bitmap mutex is taken around all metadata changes
There could be a concurrent call to bdrv_set_dirty at the same
time that bdrv_enable_dirty_bitmap is called.  Therefore, even
though only the metadata of the bitmap is changed, we need to
treat it as a write to the list and take both BQL and dirty
bitmap mutex.

Adjust the code to take the mutex whenever this kind of metadata
change happens, and adjust the comments where it is read.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>