scsi: move non-emulation specific code to scsi/
There is a bunch of SCSI code that is shared by block/iscsi.c and
hw/scsi, and the introduction of the persistent reservation helper
will add more instances of this.  There is also include/block/scsi.h,
which actually is not part of the core block layer.

Create a directory for this kind of shared code.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>